requirements-dev.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dj-static==0.0.6
Django==4.1.3
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.2
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
psycopg2-binary==2.9.5
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sentry-sdk==1.11.0
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
tomli==2.0.1
urllib3==1.26.12
